Criação de order funcionando legal, calcula o valor total e de cada item. Além de tirar do stock os que foram pedidos.

+++ Product-Service/src/main/java/br/edu/infnet/product/dto/ProductResponse.java
@@ -1,15 +1,15 @@
 package br.edu.infnet.product.dto;
 
-// ... outros imports ...
 import br.edu.infnet.product.domain.enums.Platform;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 public record ProductResponse(
         String id,
         String title,
         String description,
-        Double price,
+        BigDecimal price,
         Platform platform, // Mudou aqui
         Integer stockQuantity,
         LocalDateTime releaseDate
